refactor: extract sql predicate fragments to predicates module
